refactor(clawdstrike): extract prechecks impl from engine.rs
Move the enclave/origin/budget prechecks (enclave_mcp_precheck,
origin_data_precheck, origin_budget_precheck, consume_origin_budget) into
prechecks.rs as an impl HushEngine block, together with their free helpers
(tool_matches, origin_budget_limit, is_external_origin, the OutputSend
payload types + parser).

The precheck methods are widened to pub(crate) (called from the decision
flow that stays in mod.rs). tool_matches is re-exported under #[cfg(test)]
for test_tool_matches_helper. Public API unchanged.

Pure relocation: 825 lib tests + formal-diff-tests pass.
